Don't bitswap ABRs from signers
